Change EM100 state struct to reflect hardware
Change the types in struct em100 to match the
sizes that we read from the hardware.

Reviewed-on: http://review.coreboot.org/11252
Tested-by: build bot (Jenkins)

+++ em100.c
@@ -36,9 +36,9 @@ void exit_handler(int sig) {
 struct em100 {
 	libusb_device_handle *dev;
 	libusb_context *ctx;
-	unsigned int mcu;
-	unsigned int fpga;
-	unsigned int serialno;
+	uint16_t mcu;
+	uint16_t fpga;
+	uint32_t serialno;
 };
 
 struct em100_hold_pin_states {
